+++ src/main/java/dev/lambdaurora/lambdynlights/mixin/sodium/RSOTabAccessor.java
@@ -1,0 +1,27 @@
+/*
+ * Copyright © 2025 LambdAurora <email@lambdaurora.dev>
+ *
+ * This file is part of LambDynamicLights.
+ *
+ * Licensed under the Lambda License. For more information,
+ * see the LICENSE file.
+ */
+
+package dev.lambdaurora.lambdynlights.mixin.sodium;
+
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+import net.minecraft.network.chat.Text;
+import org.spongepowered.asm.mixin.Final;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Pseudo;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Environment(EnvType.CLIENT)
+@Pseudo
+@Mixin(targets = "me.flashyreese.mods.reeses_sodium_options.client.gui.frame.tab.Tab", remap = false)
+public interface RSOTabAccessor {
+	@Final
+	@Accessor(value = "title", remap = false)
+	Text getTitle();
+}